+++ src/Ext/SWType/Body.h
@@ -60,6 +60,7 @@ public:
 
 		std::vector<ValueableVector<int>> LimboDelivery_RandomWeightsData;
 		std::vector<ValueableVector<int>> SW_Next_RandomWeightsData;
+		std::vector<ValueableVector<int>> SW_GrantOneTime_RandomWeightsData;
 
 		std::vector<TypeConvertGroup> Convert_Pairs;
 
@@ -67,6 +68,12 @@ public:
 		Valueable<int> UseWeeds_Amount;
 		Valueable<bool> UseWeeds_StorageTimer;
 		Valueable<double> UseWeeds_ReadinessAnimationPercentage;
+
+		ValueableIdxVector<SuperWeaponTypeClass> SW_GrantOneTime;
+		Nullable<bool> SW_GrantOneTime_InitialReady;
+		ValueableVector<float> SW_GrantOneTime_RollChances;
+		Valueable<CSFText> Message_GrantOneTimeLaunched;
+		NullableIdx<VoxClass> EVA_GrantOneTimeLaunched;
 
 		ExtData(SuperWeaponTypeClass* OwnerObject) : Extension<SuperWeaponTypeClass>(OwnerObject)
 			, Money_Amount { 0 }
@@ -107,6 +114,12 @@ public:
 			, UseWeeds_Amount { RulesClass::Instance->WeedCapacity }
 			, UseWeeds_StorageTimer { false }
 			, UseWeeds_ReadinessAnimationPercentage { 0.9 }
+			, SW_GrantOneTime {}
+			, SW_GrantOneTime_InitialReady {}
+			, SW_GrantOneTime_RollChances {}
+			, SW_GrantOneTime_RandomWeightsData {}
+			, Message_GrantOneTimeLaunched {}
+			, EVA_GrantOneTimeLaunched {}
 		{ }
 
 		// Ares 0.A functions
@@ -126,6 +139,8 @@ public:
 		void ApplyDetonation(HouseClass* pHouse, const CellStruct& cell);
 		void ApplySWNext(SuperClass* pSW, const CellStruct& cell);
 		void ApplyTypeConversion(SuperClass* pSW);
+
+		void GrantOneTimeFromList(SuperClass* pSW);
 
 		virtual void LoadFromINIFile(CCINIClass* pINI) override;
 		virtual ~ExtData() = default;